pyproject.toml
[tool.poetry]
name = "converters"
version = "0.1.0"
description = "A converter for images, videos, audio and text."
authors = ["Ruzal"]

[tool.poetry.dependencies]
python = "^3.6"
Pillow = "^9.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
